# src/main/resources/openapi.yaml
openapi: 3.0.0
info:
  title: Customer Management API
  version: 1.1.0
servers:
  - url: http://localhost:8080
paths:
  /customers:
    post:
      summary: Create a new customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomerRequest'
      responses:
        '201':
          description: Customer created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerResponse'
        '400':
          description: Invalid request (e.g., email not provided or formatted incorrectly)
        '409':
          description: Duplicate email
    get:
      summary: Retrieve customers by email or name
      parameters:
        - in: query
          name: email
          schema:
            type: string
        - in: query
          name: name
          schema:
            type: string
      responses:
        '200':
          description: Customer(s) retrieved
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/CustomerResponse'
                  - type: array
                    items:
                      $ref: '#/components/schemas/CustomerResponse'
        '400':
          description: Missing or conflicting query parameters
        '404':
          description: Customer not found
  /customers/{id}:
    get:
      summary: Retrieve a customer by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Customer retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerResponse'
        '404':
          description: Customer not found
    put:
      summary: Update a customer by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomerRequest'
      responses:
        '200':
          description: Customer updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerResponse'
        '400':
          description: Invalid request
        '404':
          description: Customer not found
    delete:
      summary: Delete a customer by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Deleted
        '404':
          description: Customer not found
components:
  schemas:
    CustomerRequest:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        annualSpend:
          type: number
          format: double
        lastPurchaseDate:
          type: string
          format: date-time
      required:
        - name
        - email
    CustomerResponse:
      allOf:
        - $ref: '#/components/schemas/CustomerRequest'
        - type: object
          properties:
            id:
              type: string
            tier:
              type: string
              enum: [Silver, Gold, Platinum]